# ARROW-3958: [Plasma] Reduce number of IPCs