… cpp/src/plasma/test/client_tests.cc
… docs/source/python/plasma.rst
… python/pyarrow/tensorflow/plasma_op.cc
… python/pyarrow/tests/test_plasma.py
… python/pyarrow/tests/test_plasma_tf_op.py