@pcmoritz 2018-12-13
@robertnishihara The main rationale of this PR is to get rid of the release delay and the additional IPCs to send over the file descriptor. There is no advantage in unmapping in remapping the memory mapped files.

To be hones I would prefer to statically allocate a memory mapped file at the beginning that everybody uses, but that's not feasible with the way malloc implementations are designed (both dlmalloc and jemalloc).